Move ascii::Target to ascii::command::Target
Although it is also used by responses as well, users of the API will
primarily use if with commands. The are supplied by response but do not
need to be imported.
