+++ src/ascii/chain/setting.rs
@@ -8,10 +8,11 @@ use crate::{
 			scope::{AxisScopeMarker, DeviceScopeMarker, SameScopeAsMarker},
 			Axis, Device,
 		},
+		command::Target,
 		marker::Markers,
 		response::{check, Reply},
 		setting::Setting,
-		Port, Target,
+		Port,
 	},
 	backend::Backend,
 	error::AsciiError,
